Refactor : 디렉토리 이름 소문자 변경
