+++ src/main/java/racingcar/view/RequestMessage.java
@@ -1,4 +1,4 @@
-package racingcar.View;
+package racingcar.view;
 
 public class RequestMessage {
     private static final String CAR_NAME="경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
