+++ src/test/java/racingcar/view/ViewTest.java
@@ -1,8 +1,6 @@
-package racingcar.ViewTest;
+package racingcar.view;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
-import racingcar.View.RequestMessage;
 
 public class ViewTest {
     @Test
